Changed all unnecessary Doubles to Int

+++ src/main/java/processing_classes/Node.java
@@ -6,12 +6,12 @@ package processing_classes;
 public class Node {
 	public String name;
 	public int weight;
-	public double startTime; //TODO should we change these to double
+	public int startTime; //TODO should we change these to double
 	public int allocProc;
-	public double finishTime;
+	public int finishTime;
 	
 	//Constructor of Node class
-	public Node(String n, int w, double s, int a, double f){
+	public Node(String n, int w, int s, int a, int f){
 		name = n;
 		weight = w;
 		startTime = s;
@@ -19,11 +19,11 @@ public class Node {
 		finishTime = f;
 	}
 	
-	public void setFinish(double f){
+	public void setFinish(int f){
 		this.finishTime = f;
 	}
 	
-	public void setStart(double s){
+	public void setStart(int s){
 		this.startTime = s;
 	}
